refactor(many): extraer settings de conversación a hook — ManyPanel fase A (03/T02)
useManyConversationSettings posee los toggles tools/resources/memory/mcp,
supportsTools, userMemory, providerInfo/providerId y budgetCapApprox, con
los 3 efectos de carga (provider config + listener de cambio, mcp-enabled,
user memory) movidos verbatim. ManyPanel los consume read-only y pasa los
3 setters de toggle al composer. Refactor puro verificado en runtime
(modelo/proveedor en cabecera, toggles, indicador de contexto). 1.597→1.544.
